examples/requirements.txt
vllm
pyautogen
langchain
langchain-openai
